+++ activerecord/CHANGELOG.md
@@ -1,3 +1,13 @@
+*   Decode PostgreSQL money columns to BigDecimal instead of String when they
+    appear in direct query results.
+
+    ```ruby
+    ActiveRecord::Base.connection
+         .select_value("select '12.34'::money").class #=> BigDecimal
+    ```
+
+    *Matthew Draper*
+
 *   Add support for configuring migration strategy on a per-adapter basis.
 
     `migration_strategy` can now be set on individual adapter classes, overriding